Adding placeholder for calling search service
